For huddle 1.1.0

+++ ios/scripts/release-sdk.sh
@@ -4,25 +4,26 @@ set -e -u
 
 THIS_DIR=$(cd -P "$(dirname "$(readlink "${BASH_SOURCE[0]}" || echo "${BASH_SOURCE[0]}")")" && pwd)
 PROJECT_REPO=$(realpath ${THIS_DIR}/../..)
-RELEASE_REPO=$(realpath ${THIS_DIR}/../../../jitsi-meet-ios-sdk-releases)
 DEFAULT_SDK_VERSION=$(/usr/libexec/PlistBuddy -c "Print CFBundleShortVersionString" ${THIS_DIR}/../sdk/src/Info.plist)
 SDK_VERSION=${OVERRIDE_SDK_VERSION:-${DEFAULT_SDK_VERSION}}
 DO_GIT_TAG=${GIT_TAG:-0}
 
 
-echo "Releasing Jitsi Meet SDK ${SDK_VERSION}"
+if [ "$N2P_JITSI_RELEASE_REPO" ]; then
+	echo "Releasing Jitsi Meet SDK ${SDK_VERSION}"
 
-${THIS_DIR}/../../node_modules/react-native-webrtc/tools/downloadBitcode.sh
+    ${THIS_DIR}/../../node_modules/react-native-webrtc/tools/downloadBitcode.sh
 
-pushd ${RELEASE_REPO}
+	pushd ${N2P_JITSI_RELEASE_REPO}/ios
 
-# Generate podspec file
-cat JitsiMeetSDK.podspec.tpl | sed -e s/VERSION/${SDK_VERSION}/g > JitsiMeetSDK.podspec
+	# Generate podspec file
+	cat JitsiMeetSDK.podspec.tpl | sed -e s/VERSION/${SDK_VERSION}/g > JitsiMeetSDK.podspec
 
-# Cleanup
-rm -rf Frameworks/*
+	# Cleanup
+	rm -rf Frameworks/*
 
-popd
+	popd
+fi
 
 # Build the SDK
 pushd ${PROJECT_REPO}
@@ -59,19 +60,22 @@ if [[ $DO_GIT_TAG == 1 ]]; then
 fi
 popd
 
-pushd ${RELEASE_REPO}
+if [ "$N2P_JITSI_RELEASE_REPO" ]; then
+	pushd ${N2P_JITSI_RELEASE_REPO}/ios
 
-# Put the new files in the repo
-cp -a ${PROJECT_REPO}/ios/sdk/out/JitsiMeetSDK.xcframework Frameworks/
+	# Put the new files in the repo
+	cp -a ${PROJECT_REPO}/ios/sdk/out/JitsiMeetSDK.xcframework Frameworks/
+	#cp -a ${PROJECT_REPO}/node_modules/react-native-webrtc/apple/WebRTC.xcframework Frameworks/
 
-# Add all files to git
-if [[ $DO_GIT_TAG == 1 ]]; then
-    git add -A .
-    git commit -m "${SDK_VERSION}"
-    git tag ${SDK_VERSION}
+	# Add all files to git
+	if [[ $DO_GIT_TAG == 1 ]]; then
+	    git add -A .
+	    git commit -m "${SDK_VERSION}"
+	    git tag ${SDK_VERSION}
+	fi
+
+	popd
 fi
-
-popd
 
 echo "Finished! Don't forget to push the tags and releases repo artifacts."
 echo "The new pod can be pushed to CocoaPods by doing: pod trunk push JitsiMeetSDK.podspec"
